Dynamic graph of besier points is now working.
This was done using an updater thread. The problem was when the updater thread drew the plot, it got stuck in an infinite loop. Therefore I created an updater thread when ever the old updater thread detected a change. When the old updater thread detected a change, it would update the GUI and kick off the new updater thread which will now be the old updater thread. This thread will in turn create another updater thread and so on. The program always levels back down to 3 threads.